requirements.txt
discord.py==2.3.2
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.12.0
python-dotenv==1.0.0
aiohttp==3.9.1
SQLAlchemy==2.0.23
google-generativeai==0.3.1
pynacl==1.5.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-logging-loki==0.3.1
prometheus-client==0.19.0
asyncio==3.4.3
aiosqlite==0.19.0 